Backend/requirements.txt
Flask
requests
beautifulsoup4
langchain
google-generativeai
python-dotenv
